move member handling functions into memberService
