+++ src/api/supabase.tsx
@@ -1,3 +1,4 @@
+
 import { createClient } from "@supabase/supabase-js";
 
 const supabaseUrl = "https://anlhkcaoszuqiatipxjr.supabase.co";
